Adicionado o começo da parte de transação
Iniciei a parte de transação, adicionando os primeiros comandos na main pra testar se tava chegando os valores, guardando em txt certinho, indo agora para a parte adicional, que entra o calculo de saldo total, despesas totais, entre outros métodos.

+++ gerenciadorFinanceiro/src/Transacao.java
@@ -3,12 +3,12 @@ import java.time.LocalDate;
 public class Transacao {
 
     private final int id;
-    private int valor;
+    private double valor;
     private int idCategoria;
     private LocalDate data;
     private String tipo; // "receita" ou "despesa"
 
-    public Transacao(int id, int valor, int idCategoria, LocalDate data, String tipo) {
+    public Transacao(int id, double valor, int idCategoria, LocalDate data, String tipo) {
         this.id = id;
         this.valor = valor;
         this.idCategoria = idCategoria;
